# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! Directly
    from your PC, you can upload your photos, and they will be visible to everyone who is following you.
  version: "0.0.1"
...
paths:
...
  /profile/:
    get:
      tags: ["Interactions with other users"]
      summary: Search users
      description: |
        Retrieves all users that match the username passed in the query. If the feedPageId is not passed, the first page will returned.
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - {$ref: '#/components/parameters/userToSearch'}
        - {$ref: '#/components/parameters/pageId'}
      responses:
        '200':
          description: |
            Obtain the list of the user. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users'
        '400':  {$ref: '#/components/responses/badRequest'}
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '5XX':  {$ref: '#/components/responses/serverError'}
...
  /profile/{username}/photos/:  
    parameters:
      - {$ref: '#/components/parameters/username'}
    post:
      tags: ["Profile operations"]
      summary: Upload a new photo
      description: |
        Lets user to upload a new photo in their profile.
      operationId: uploadPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: New photo
        content:
          image/*:
            schema:
              $ref: '#/components/schemas/file'          
        required: true
      responses:
        '201':
          description: |
            File uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '415':  {$ref: '#/components/responses/fileNotSupported'}
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '5XX':  {$ref: '#/components/responses/serverError'}
  /profile/{username}/photos/{photoId}:  
    parameters:
      - {$ref: '#/components/parameters/username'}
      - {$ref: '#/components/parameters/photoId'}
    delete:
      tags: ["Profile operations"]
      summary: Delete a photo
      description: |
        Lets user to delete a photo.
      operationId: deletePhoto
      security:
        - bearerAuth: []
      responses:
        '204':
          description: |
            File deleted successfully.
        '400':  {$ref: '#/components/responses/badRequest'}
        '404':  {$ref: '#/components/responses/notFoundError'}  
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '5XX':  {$ref: '#/components/responses/serverError'}
  /profile/{username}/banned/{byUsername}:
    parameters:
      - {$ref: '#/components/parameters/username'}
      - {$ref: '#/components/parameters/byUsername'}
    put:
      tags: ["Interactions with other users"]
      summary: Ban another user
      description: |
        Lets user to ban another user.
      operationId: banUser
      security:
        - bearerAuth: []
      responses:
        '201':
          description: |
            User banned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '404':  {$ref: '#/components/responses/notFoundError'}  
        '409':  {$ref: '#/components/responses/conflictError'}
        '5XX':  {$ref: '#/components/responses/serverError'}
    delete:
      tags: ["Interactions with other users"]
      summary: Unban another user
      description: |
        Lets user to unban another user.
      operationId: unbanUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: |
            User unbanned successfully.
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '404':  {$ref: '#/components/responses/notFoundError'}  
        '409':  {$ref: '#/components/responses/conflictError'}
        '5XX':  {$ref: '#/components/responses/serverError'}
  /profile/{username}/followed/{byUsername}:
    parameters:
      - {$ref: '#/components/parameters/username'}
      - {$ref: '#/components/parameters/byUsername'}
    put:
      tags: ["Interactions with other users"]
      summary: Follow a user
      description: |
        Lets user to follow another user.
      operationId: followUser
      security:
        - bearerAuth: []
      responses:
        '201':
          description: |
            User followed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '404':  {$ref: '#/components/responses/notFoundError'}  
        '409':  {$ref: '#/components/responses/conflictError'}
        '5XX':  {$ref: '#/components/responses/serverError'}
    delete:
      tags: ["Interactions with other users"]
      summary: Unfollow a user
      description: |
        Lets user to unfollow another user.
      operationId: unfollowUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: |
            User unfollowed successfully.
        '404':  {$ref: '#/components/responses/notFoundError'}  
        '409':  {$ref: '#/components/responses/conflictError'}
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '5XX':  {$ref: '#/components/responses/serverError'}
  /feed/:
    get:
      tags: ["Feed operations"]
      summary: Retrive photo stream
      description: |
        Retrieves all the photo shared by your following. If the pageId is not passed, the first page will returned.
      operationId: getMyStream
      security:
        - bearerAuth: []
      parameters:
        - {$ref: '#/components/parameters/pageId'}
      responses:
        '200':
          description: |
            Obtain the list of image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/feedPhoto'
        '204':
          description: |
            There isn't any image to show.
        '400':  {$ref: '#/components/responses/badRequest'}
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '5XX':  {$ref: '#/components/responses/serverError'}
  /feed/{photoId}/: 
    parameters:
      - {$ref: '#/components/parameters/photoId'}
    get:
      tags: ["Feed operations"]
      summary: Retrive a photo.
      description: |
        Retrieves a photo that match with the photoId passed in input.
      operationId: getPhoto
      security:
        - bearerAuth: []
      responses:
        '200':
          description: |
            Obtain an image.
          content:
            image/*:
              schema:
                $ref: '#/components/schemas/file'
        '400':  {$ref: '#/components/responses/badRequest'}
        '404':  {$ref: '#/components/responses/notFoundError'}
        '401':  {$ref: '#/components/responses/unauthAccess'}
        '5XX':  {$ref: '#/components/responses/serverError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    username: 
      description: User's name.
      name: username
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/stringUsername'
    byUsername:
      description: User that do the action.
      name: byUsername
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/stringUsername'
    photoId:
      description: Photo identifier.
      name: photoId
      in: path
      required: true
      schema:
        type: integer
    userToSearch:
      description: User to search.
      name: userToSearch
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/stringUsername'
    pageId:
      description: Page identifier.
      name: pageId
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
  schemas:  
    stringUsername:
      title: Username info
      type: string
      pattern: '^[a-zA-Z0-9_.]*$' 
      minLength: 3
      maxLength: 16
      example: Maria
      description: User's username.
    stringIdentifier:
      title: User identifier info
      type: string
      minLength: 36
      maxLength: 36
      example: 123e4567-e89b-12d3-a456-426614174000
      description: User identifier (uuid).
    integerPhoto:
      title: Photo identifier info
      type: integer
      example: 1452
      minimum: 0
      description: Photo identifier.
...
    feedPhoto:
      title: Feed containing photos
      description: |
        Represents the feed.
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/postsExample'
          minItems: 1
          maxItems: 10
          description: List of photos.
        nextFeedPageId:
          type: integer
          minimum: 1
          example: 7
          description: Id of the next page of posts.
    postsExample:
      title: Post example
      description: |
        Represents one post.
      type: object
      properties:
        photo:  
          $ref: '#/components/schemas/integerPhoto'
        data:
          type: string
          format: date-time
          description: Upload date
          example: "2022-01-01T00:00:00Z"
        user:
          $ref: '#/components/schemas/stringIdentifier'
        numberOfLikes:
          type: integer
          example: 540
          minimum: 0
          description: Number of likes.
        numberOfComments:
          type: integer
          example: 15
          minimum: 0
          description: Number of comments.
...
    user:  
      title: User
      description: Represents an identifier and a username for one user.
      type: object
      properties:
        identifier:
          $ref: '#/components/schemas/stringIdentifier'
        username:
          $ref: '#/components/schemas/stringUsername'
    users:
      title: List of users
      description: |
        Represents an object that includes a list of 10 user.
      type: object
      properties: 
        users:
          type: array
          items:
            $ref: '#/components/schemas/user'
          minItems: 1
          maxItems: 10
          description: List of Users.
        nextUsersPageId: 
          type: integer
          minimum: 1
          example: 7
          description: Id of the next page of users.
    success:
      title: Success message
      description: |
        Represents an object that includes a message.
      type: object
      properties:
        message:
          type: string
          example: Resource created
          description: Success message.
...
    file:
      title: Image
      description: Represents an image.
      type: string
      format: binary   
      minLength: 0
      maxLength: 8388608
    errorMessage:
      title: Error message.
      description: Represents the error message.
      type: object
      properties:
        message: 
          type: string
          pattern: '^[a-zA-Z0-9_., !?:;""$%&€()[{}]*$'
          example: Error message.
          minLength: 0
          maxLength: 50
          description: Represents the text of the error.
  responses:
    badRequest:
      description: Bad request. 
      content:
        application/json:
          schema:  
            $ref: '#/components/schemas/errorMessage'
    unauthAccess:
      description: Unauthorized access.
      content:
        application/json:
          schema:  
            $ref: '#/components/schemas/errorMessage'
...
    conflictError:
      description: Conflict error. 
      content:
        application/json:
          schema:  
            $ref: '#/components/schemas/errorMessage'
    serverError:
      description: Internal server error. 
      content:
        application/json:
          schema:  
            $ref: '#/components/schemas/errorMessage'
    notFoundError:
      description: Not found. 
      content:
        application/json:
          schema:  
            $ref: '#/components/schemas/errorMessage'
    fileNotSupported:
      description: File format not supported.
      content:
        application/json:
          schema:  
            $ref: '#/components/schemas/errorMessage'